more tests to make ci leaner, add cleanup to modal ci
